code refactored
