+++ src/types/LocationData.ts
@@ -1,4 +1,10 @@
 export type LocationData = {
   name: string;
+  region: string;
   country: string;
+  lat: number;
+  lon: number;
+  tz_id: string;
+  localtime_epoch: number;
+  localtime: string;
 };
